# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "plant-assistant",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/plants/": {
      "post": {
        "tags": [
          "plants"
        ],
        "summary": "Create Plant",
        "description": "Create a new plant for the current user.",
        "operationId": "create_plant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "plants"
        ],
        "summary": "Get Plants",
        "description": "Get paginated list of user's plants.",
        "operationId": "get_plants",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Search"
            }
          },
          {
            "name": "species_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Species Id"
            }
          },
          {
            "name": "is_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Is Archived"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/{plant_id}": {
      "get": {
        "tags": [
          "plants"
        ],
        "summary": "Get Plant",
        "description": "Get a specific plant by ID.",
        "operationId": "get_plant",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "plants"
        ],
        "summary": "Update Plant",
        "description": "Update a plant.",
        "operationId": "update_plant",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plants"
        ],
        "summary": "Delete Plant",
        "description": "Delete (archive) a plant.",
        "operationId": "delete_plant",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/identify/": {
      "post": {
        "tags": [
          "plant-identification"
        ],
        "summary": "Identify Plant",
        "description": "Identify a plant from images and/or description.",
        "operationId": "identify_plant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plants/care/advice": {
      "post": {
        "tags": [
          "plant-care"
        ],
        "summary": "Get Care Advice",
        "description": "Get personalized care advice for a plant.",
        "operationId": "get_care_advice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/diagnose/": {
      "post": {
        "tags": [
          "plant-diagnosis"
        ],
        "summary": "Diagnose Plant",
        "description": "Diagnose plant issues from symptoms and/or images.",
        "operationId": "diagnose_plant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnoseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnoseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plants/chat/": {
      "post": {
        "tags": [
          "plant-chat"
        ],
        "summary": "Chat With Assistant",
        "description": "Chat with the plant care assistant.",
        "operationId": "chat_with_assistant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CareRequest": {
        "properties": {
          "plant_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Id"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "description": "ZIP code or location"
          },
          "environment": {
            "additionalProperties": true,
            "type": "object",
            "title": "Environment",
            "description": "Environment details"
          },
          "preferences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferences",
            "description": "User care preferences"
          }
        },
        "type": "object",
        "required": [
          "location",
          "environment"
        ],
        "title": "CareRequest",
        "description": "Schema for plant care advice request."
      },
      "CareResponse": {
        "properties": {
          "plan": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Plan",
            "description": "Care instructions by category (watering, light, soil, etc.)"
          },
          "seasonal_adjustments": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Seasonal Adjustments",
            "description": "Seasonal care modifications"
          },
          "eco_tips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Eco Tips",
            "description": "Environmentally friendly care suggestions"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "Reference sources for recommendations"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "Care advice disclaimer"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "seasonal_adjustments",
          "eco_tips",
          "sources",
          "disclaimer"
        ],
        "title": "CareResponse",
        "description": "Schema for plant care advice response."
      },
...
      "DiagnoseRequest": {
        "properties": {
          "symptoms": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Symptoms"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Images"
          },
          "plant_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plant Id"
          }
        },
        "type": "object",
        "title": "DiagnoseRequest",
        "description": "Schema for plant diagnosis request."
      },
      "DiagnoseResponse": {
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/PlantIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "remedies": {
            "items": {
              "$ref": "#/components/schemas/Remedy"
            },
            "type": "array",
            "title": "Remedies"
          },
          "prevention": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prevention"
          },
          "similar_cases": {
            "type": "integer",
            "title": "Similar Cases",
            "description": "Number of similar cases found"
          },
          "severity": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Severity",
            "description": "Overall severity rating"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "issues",
          "remedies",
          "prevention",
          "similar_cases",
          "severity",
          "disclaimer"
        ],
        "title": "DiagnoseResponse",
        "description": "Schema for plant diagnosis response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdentificationConfidence": {
        "type": "string",
        "enum": [
          "very_low",
          "low",
          "medium",
          "high",
          "very_high"
        ],
        "title": "IdentificationConfidence",
        "description": "Confidence levels for plant identification."
      },
      "IdentifyRequest": {
        "properties": {
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Images",
            "description": "Base64-encoded images"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 750,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Text description of the plant"
          },
          "user_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Metadata",
            "description": "Additional context like location, conditions"
          }
        },
        "type": "object",
        "title": "IdentifyRequest",
        "description": "Schema for plant identification request."
      },
      "IdentifyResponse": {
        "properties": {
          "primary_identification": {
            "$ref": "#/components/schemas/PlantIdentification"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/PlantIdentification"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "fun_facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fun Facts",
            "default": []
          },
          "basic_info": {
            "type": "string",
            "title": "Basic Info"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms"
          },
          "confidence_level": {
            "$ref": "#/components/schemas/IdentificationConfidence"
          }
        },
        "type": "object",
        "required": [
          "primary_identification",
          "basic_info",
          "disclaimer",
          "processing_time_ms",
          "confidence_level"
        ],
        "title": "IdentifyResponse",
        "description": "Schema for plant identification response."
      },
...
      "PlantCreate": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "species_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species Id"
          },
          "acquisition_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquisition Date"
          }
        },
        "type": "object",
        "title": "PlantCreate",
        "description": "Schema for creating a new plant."
      },
...
      "PlantListResponse": {
        "properties": {
          "plants": {
            "items": {
              "$ref": "#/components/schemas/PlantResponse"
            },
            "type": "array",
            "title": "Plants"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "plants",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "PlantListResponse",
        "description": "Schema for paginated plant list response."
      },
...
      "PlantResponse": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "species_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species Id"
          },
          "acquisition_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquisition Date"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          },
          "ai_metrics_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Metrics Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/PlantStatus",
            "default": "unknown"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count",
            "default": 0
          },
          "last_care_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Care Date"
          },
          "next_reminder_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Reminder Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "species_id",
          "acquisition_date",
          "is_archived",
          "ai_metrics_json",
          "created_at",
          "updated_at"
        ],
        "title": "PlantResponse",
        "description": "Schema for plant response with additional computed fields."
      },
      "PlantStatus": {
        "type": "string",
        "enum": [
          "healthy",
          "warning",
          "critical",
          "unknown"
        ],
        "title": "PlantStatus",
        "description": "Status of a plant."
      },
      "PlantUpdate": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "species_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species Id"
          },
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          }
        },
        "type": "object",
        "title": "PlantUpdate",
        "description": "Schema for updating a plant."
      },
...
    }
  }
}